Add files via upload
To run this project, make a new flutter project, with nothing in it. Then drag in all of the .dart files into the folder called lib, and replace the default pubsepc.yaml file with the attached .yaml file. Finally, find the folder named ios, and replace the file named "Runner.xcworkspace" with the .xcworkspace file attached. 

The pill.py is the server backend, so this should be ran before the app is ran through an IOS simulator.